+++ src/simmate/calculators/vasp/database/energy.py
@@ -11,6 +11,14 @@ class MatprojStaticEnergy(StaticEnergy):
     pass
 
 
+class MatprojHSEStaticEnergy(StaticEnergy):
+    pass
+
+
+class MatprojSCANStaticEnergy(StaticEnergy):
+    pass
+
+
 class NEBEndpointStaticEnergy(StaticEnergy):
     pass