+++ src/simmate/calculators/vasp/tasks/static_energy/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from .materials_project import MatprojStaticEnergy
+from .materials_project_scan import MatprojSCANStaticEnergy
 from .mit import MITStaticEnergy
 from .quality_04 import Quality04Energy
 from .neb_endpoint import NEBEndpointStaticEnergy
+from .materials_project_hse import MatprojHSEStaticEnergy