+++ src/simmate/calculators/vasp/workflows/static_energy/all.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from .matproj import StaticEnergy__Vasp__Matproj
+from .matproj_hse import StaticEnergy__Vasp__MatprojHse
+from .matproj_scan import StaticEnergy__Vasp__MatprojScan
 from .mit import StaticEnergy__Vasp__Mit
 from .quality04 import StaticEnergy__Vasp__Quality04
 from .neb_endpoint import StaticEnergy__Vasp__NebEndpoint