+++ src/simmate/database/workflow_results/__init__.py
@@ -7,6 +7,10 @@ __doc__ = get_doc_from_readme(__file__)
 from simmate.calculators.vasp.database.relaxation import (
     MatprojRelaxation,
     MatprojIonicStep,
+    MatprojHSERelaxation,
+    MatprojHSEIonicStep,
+    MatprojSCANRelaxation,
+    MatprojSCANIonicStep,
     MITRelaxation,
     MITIonicStep,
     Quality00Relaxation,
@@ -26,6 +30,8 @@ from simmate.calculators.vasp.database.relaxation import (
 
 from simmate.calculators.vasp.database.energy import (
     MatprojStaticEnergy,
+    MatprojHSEStaticEnergy,
+    MatprojSCANStaticEnergy,
     MITStaticEnergy,
     Quality04StaticEnergy,
     NEBEndpointStaticEnergy,
@@ -33,10 +39,12 @@ from simmate.calculators.vasp.database.energy import (
 
 from simmate.calculators.vasp.database.band_structure import (
     MatprojBandStructure,
+    MatprojHSEBandStructure,
 )
 
 from simmate.calculators.vasp.database.density_of_states import (
     MatprojDensityOfStates,
+    MatprojHSEDensityOfStates,
 )
 
 from simmate.calculators.vasp.database.nudged_elastic_band import (
@@ -48,6 +56,10 @@ from simmate.calculators.vasp.database.nudged_elastic_band import (
 from simmate.calculators.vasp.database.dynamics import (
     MITDynamicsRun,
     MITDynamicsIonicStep,
+    MatprojDynamicsRun,
+    MatprojDynamicsIonicStep,
+    MatVirtualLabNPTDynamicsRun,
+    MatVirtualLabNPTDynamicsIonicStep,
 )
 
 from simmate.calculators.vasp.database.population_analysis import (